+++ src/Engine/Camera/Camera.h
@@ -47,7 +47,7 @@ class Camera {
         [[nodiscard]] bool getMouseCapture() const;
         void toggleMouseCapture();
 
-        void setFOV(float _fov);
+        void setFOV(uint8_t _fov);
         void setPosition(glm::vec3 newPos);
         [[nodiscard]] glm::vec3 getPosition() const;
         [[nodiscard]] glm::vec2 getRotation() const;
